kruft removal, some build errors remain
