+++ zingolib/src/wallet/output.rs
@@ -20,17 +20,7 @@ use super::error::WalletError;
 use super::transaction::transaction_unspent_outputs;
 use super::LightWallet;
 
-pub use interface::OutputInterface as OldOutputInterface;
-pub use interface::ShieldedNoteInterface;
-pub use orchard::OrchardNote;
-pub use sapling::SaplingNote;
-pub use transparent::TransparentOutput;
-
-pub mod interface;
-pub mod orchard;
 pub mod query;
-pub mod sapling;
-pub mod transparent;
 
 /// Spend status of an output
 #[derive(Clone, Copy, PartialEq, Debug)]
@@ -314,70 +304,6 @@ impl LightWallet {
     }
 }
 
-/// An interface for accessing all the common functionality of all the outputs
-#[enum_dispatch::enum_dispatch(OldOutputInterface)]
-#[non_exhaustive] // We can add new pools later
-#[derive(Clone, Debug)]
-pub enum Output {
-    /// Transparent Outputs
-    TransparentOutput,
-    /// Sapling Notes
-    SaplingNote,
-    /// Orchard Notes
-    OrchardNote,
-}
-impl Output {
-    /// All the output records
-    pub fn get_record_outputs(
-        transaction_record: &super::transaction_record::TransactionRecord,
-    ) -> Vec<Self> {
-        transaction_record
-            .transparent_outputs
-            .iter()
-            .map(|output| Self::TransparentOutput(output.clone()))
-            .chain(
-                transaction_record
-                    .sapling_notes
-                    .iter()
-                    .map(|output| Self::SaplingNote(output.clone())),
-            )
-            .chain(
-                transaction_record
-                    .orchard_notes
-                    .iter()
-                    .map(|output| Self::OrchardNote(output.clone())),
-            )
-            .collect()
-    }
-
-    /// Every notes' outputinterface for a given spend status
-    pub fn get_all_outputs_with_status(
-        transaction_record: &super::transaction_record::TransactionRecord,
-        spend_status_query: OutputSpendStatusQuery,
-    ) -> Vec<Self> {
-        transaction_record
-            .transparent_outputs
-            .iter()
-            .filter(|output| output.spend_status_query(spend_status_query))
-            .map(|output| Self::TransparentOutput(output.clone()))
-            .chain(
-                transaction_record
-                    .sapling_notes
-                    .iter()
-                    .filter(|output| output.spend_status_query(spend_status_query))
-                    .map(|output| Self::SaplingNote(output.clone())),
-            )
-            .chain(
-                transaction_record
-                    .orchard_notes
-                    .iter()
-                    .filter(|output| output.spend_status_query(spend_status_query))
-                    .map(|output| Self::OrchardNote(output.clone())),
-            )
-            .collect()
-    }
-}
-
 pub(crate) enum RemainingNeeded {
     Positive(NonNegativeAmount),
     GracelessChangeAmount(NonNegativeAmount),
@@ -410,6 +336,8 @@ fn calculate_remaining_needed(
     }
 }
 
+// FIXME: zingo2, update for new output types
+/*
 #[cfg(test)]
 pub mod mocks {
     //! Mock version of the struct for testing
@@ -582,3 +510,4 @@ pub mod tests {
         assert!(sapling_spent_note.query(spent_any_pool_query));
     }
 }
+*/
